[FIX] l10n_ar_afipws_fe: Refactor of account_move and other tweaks
